def get_call_info_id2fname(cls, db, QC_method_id, call_method_id, filter_calls_QCed=1, max_call_info_mismatch_rate=1, debug=0, min_no_of_non_NA_pairs=40, **keywords): 	""" 	2008-09-16 		add **keywords in the argument 		deal with input_dir if it's present. replace the directory in call_info.filename with it.
change the meaning of input_dir. now if it's a given directory, replace the filename directory fetched from the database
('input_dir', 0, ): [None, 'n', 1, 'If given is directory, call_info.filename is assumed to be in this directory. If it is a FILE, call data would be read from it.'],\

git-svn-id: file:///usr/local/svn/repos@1122 c5b91a9c-90b7-4db6-b2ac-9ef0fe68db40