from Roger:
Here is an example of how to modify the script to read from a text file and create the binary file. 
But it will be much more efficient to do it with a separate script line by line instead of having to load the entire dataset into memory. 

The algorithm uses a binary file that is mapped directly into memory. This seems to work well on very large datasets. 


git-svn-id: file:///usr/local/svn/repos@1759 c5b91a9c-90b7-4db6-b2ac-9ef0fe68db40